Move CL tRPC router to local (#2947)
* move CL tRPC router to local

* remove CL edge

* move swap to local
