+++ packages/web/components/cards/my-position/expanded.tsx
@@ -560,7 +560,7 @@ const Chart: FunctionComponent<{
 });
 
 const SuperfluidPositionInfo: FunctionComponent<
-  RouterOutputs["edge"]["concentratedLiquidity"]["getPositionDetails"]["superfluidData"]
+  RouterOutputs["local"]["concentratedLiquidity"]["getPositionDetails"]["superfluidData"]
 > = (props) => {
   const {
     validatorName,
